Repository, Mapper 테스트코드 위치 수정완료
